refactor     extract TagCloudHandler from ConsoleUI
