+++ TagCloudUI/Infrastructure/ITagCloudHandler.cs
@@ -1,0 +1,10 @@
+﻿using System.Drawing;
+using TagCloud.Core;
+
+namespace TagCloudUI.Infrastructure
+{
+    public interface ITagCloudHandler
+    {
+        Result<Bitmap> Run(IAppSettings settings);
+    }
+}
